+++ geopackage-map/src/main/java/mil/nga/geopackage/map/features/FeatureInfoBuilder.java
@@ -3,7 +3,9 @@ package mil.nga.geopackage.map.features;
 import android.content.Context;
 import android.content.res.Resources;
 import android.util.Log;
+import android.view.View;
 
+import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.LatLng;
 
 import java.sql.SQLException;
@@ -15,6 +17,10 @@ import java.util.Map;
 import java.util.Set;
 
 import mil.nga.geopackage.GeoPackageException;
+import mil.nga.geopackage.extension.nga.style.FeatureStyleExtension;
+import mil.nga.geopackage.extension.nga.style.FeatureTableStyles;
+import mil.nga.geopackage.extension.nga.style.IconRow;
+import mil.nga.geopackage.extension.nga.style.StyleRow;
 import mil.nga.geopackage.extension.schema.columns.DataColumns;
 import mil.nga.geopackage.extension.schema.columns.DataColumnsDao;
 import mil.nga.geopackage.features.index.FeatureIndexListResults;
@@ -29,6 +35,7 @@ import mil.nga.geopackage.map.geom.GoogleMapShapeConverter;
 import mil.nga.geopackage.map.tiles.overlay.FeatureOverlayQuery;
 import mil.nga.geopackage.srs.SpatialReferenceSystem;
 import mil.nga.geopackage.srs.SpatialReferenceSystemDao;
+import mil.nga.geopackage.style.PixelBounds;
 import mil.nga.geopackage.tiles.overlay.FeatureRowData;
 import mil.nga.geopackage.tiles.overlay.FeatureTableData;
 import mil.nga.proj.Projection;
@@ -50,6 +57,11 @@ public class FeatureInfoBuilder {
      * Feature DAO
      */
     private final FeatureDao featureDao;
+
+    /**
+     * Feature Style Extension
+     */
+    private FeatureTableStyles featureStyles;
 
     /**
      * Geometry Type
@@ -98,8 +110,21 @@ public class FeatureInfoBuilder {
      * @param featureDao feature dao
      */
     public FeatureInfoBuilder(Context context, FeatureDao featureDao) {
+        this(context, featureDao, null);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context       context
+     * @param featureDao    feature dao
+     * @param featureStyles feature table styles
+     * @since 6.3.0
+     */
+    public FeatureInfoBuilder(Context context, FeatureDao featureDao, FeatureTableStyles featureStyles) {
 
         this.featureDao = featureDao;
+        this.featureStyles = featureStyles;
 
         geometryType = featureDao.getGeometryType();
         name = featureDao.getDatabase() + " - " + featureDao.getTableName();
@@ -237,7 +262,7 @@ public class FeatureInfoBuilder {
      * @return results message or null if no results
      */
     public String buildResultsInfoMessageAndClose(FeatureIndexResults results) {
-        return buildResultsInfoMessageAndClose(results, 0, null, null);
+        return buildResultsInfoMessageAndClose(results, 0.0, null, null);
     }
 
     /**
@@ -248,7 +273,7 @@ public class FeatureInfoBuilder {
      * @return results message or null if no results
      */
     public String buildResultsInfoMessageAndClose(FeatureIndexResults results, Projection projection) {
-        return buildResultsInfoMessageAndClose(results, 0, null, projection);
+        return buildResultsInfoMessageAndClose(results, 0.0, null, projection);
     }
 
     /**
@@ -277,6 +302,51 @@ public class FeatureInfoBuilder {
 
         try {
             message = buildResultsInfoMessage(results, tolerance, clickLocation, projection);
+        } finally {
+            results.close();
+        }
+
+        return message;
+    }
+
+    /**
+     * Build a feature results information message and close the results
+     *
+     * @param results               feature index results
+     * @param tolerance             distance tolerance
+     * @param clickLocation         map click location
+     * @param density               display density: {@link android.util.DisplayMetrics#density}
+     * @param zoom                  current zoom level
+     * @param view                  view
+     * @param map                   Google Map
+     * @param screenClickPercentage screen click percentage between 0.0 and 1.0
+     * @return results message or null if no results
+     * @since 6.3.0
+     */
+    public String buildResultsInfoMessageAndClose(FeatureIndexResults results, double tolerance, LatLng clickLocation, float density, double zoom, View view, GoogleMap map, float screenClickPercentage) {
+        return buildResultsInfoMessageAndClose(results, tolerance, clickLocation, density, zoom, view, map, screenClickPercentage, null);
+    }
+
+    /**
+     * Build a feature results information message and close the results
+     *
+     * @param results               feature index results
+     * @param tolerance             distance tolerance
+     * @param clickLocation         map click location
+     * @param density               display density: {@link android.util.DisplayMetrics#density}
+     * @param zoom                  current zoom level
+     * @param view                  view
+     * @param map                   Google Map
+     * @param screenClickPercentage screen click percentage between 0.0 and 1.0
+     * @param projection            desired geometry projection
+     * @return results message or null if no results
+     * @since 6.3.0
+     */
+    public String buildResultsInfoMessageAndClose(FeatureIndexResults results, double tolerance, LatLng clickLocation, float density, double zoom, View view, GoogleMap map, float screenClickPercentage, Projection projection) {
+        String message = null;
+
+        try {
+            message = buildResultsInfoMessage(results, tolerance, clickLocation, density, zoom, view, map, screenClickPercentage, projection);
         } finally {
             results.close();
         }
@@ -329,11 +399,48 @@ public class FeatureInfoBuilder {
      * @return results message or null if no results
      */
     public String buildResultsInfoMessage(FeatureIndexResults results, double tolerance, LatLng clickLocation, Projection projection) {
+        return buildResultsInfoMessage(results, tolerance, clickLocation, 1.0f, 0.0, null, null, 0.0f, projection);
+    }
+
+    /**
+     * Build a feature results information message
+     *
+     * @param results               feature index results
+     * @param tolerance             distance tolerance
+     * @param clickLocation         map click location
+     * @param density               display density: {@link android.util.DisplayMetrics#density}
+     * @param zoom                  current zoom level
+     * @param view                  view
+     * @param map                   Google Map
+     * @param screenClickPercentage screen click percentage between 0.0 and 1.0
+     * @return results message or null if no results
+     * @since 6.3.0
+     */
+    public String buildResultsInfoMessage(FeatureIndexResults results, double tolerance, LatLng clickLocation, float density, double zoom, View view, GoogleMap map, float screenClickPercentage) {
+        return buildResultsInfoMessage(results, tolerance, clickLocation, density, zoom, view, map, screenClickPercentage, null);
+    }
+
+    /**
+     * Build a feature results information message
+     *
+     * @param results               feature index results
+     * @param tolerance             distance tolerance
+     * @param clickLocation         map click location
+     * @param density               display density: {@link android.util.DisplayMetrics#density}
+     * @param zoom                  current zoom level
+     * @param view                  view
+     * @param map                   Google Map
+     * @param screenClickPercentage screen click percentage between 0.0 and 1.0
+     * @param projection            desired geometry projection
+     * @return results message or null if no results
+     * @since 6.3.0
+     */
+    public String buildResultsInfoMessage(FeatureIndexResults results, double tolerance, LatLng clickLocation, float density, double zoom, View view, GoogleMap map, float screenClickPercentage, Projection projection) {
 
         String message = null;
 
         // Fine filter results so that the click location is within the tolerance of each feature row result
-        FeatureIndexResults filteredResults = fineFilterResults(results, tolerance, clickLocation);
+        FeatureIndexResults filteredResults = fineFilterResults(results, tolerance, clickLocation, density, zoom, view, map, screenClickPercentage);
 
         long featureCount = filteredResults.count();
         if (featureCount > 0) {
@@ -454,11 +561,48 @@ public class FeatureInfoBuilder {
      * @return feature table data or null if not results
      */
     public FeatureTableData buildTableDataAndClose(FeatureIndexResults results, double tolerance, LatLng clickLocation, Projection projection) {
+        return buildTableDataAndClose(results, tolerance, clickLocation, 1.0f, 0.0, null, null, 0.0f, projection);
+    }
+
+    /**
+     * Build a feature results information message
+     *
+     * @param results               feature index results
+     * @param tolerance             distance tolerance
+     * @param clickLocation         map click location
+     * @param density               display density: {@link android.util.DisplayMetrics#density}
+     * @param zoom                  current zoom level
+     * @param view                  view
+     * @param map                   Google Map
+     * @param screenClickPercentage screen click percentage between 0.0 and 1.0
+     * @return feature table data or null if not results
+     * @since 6.3.0
+     */
+    public FeatureTableData buildTableDataAndClose(FeatureIndexResults results, double tolerance, LatLng clickLocation, float density, double zoom, View view, GoogleMap map, float screenClickPercentage) {
+        return buildTableDataAndClose(results, tolerance, clickLocation, density, zoom, view, map, screenClickPercentage, null);
+    }
+
+    /**
+     * Build a feature results information message
+     *
+     * @param results               feature index results
+     * @param tolerance             distance tolerance
+     * @param clickLocation         map click location
+     * @param density               display density: {@link android.util.DisplayMetrics#density}
+     * @param zoom                  current zoom level
+     * @param view                  view
+     * @param map                   Google Map
+     * @param screenClickPercentage screen click percentage between 0.0 and 1.0
+     * @param projection            desired geometry projection
+     * @return feature table data or null if not results
+     * @since 6.3.0
+     */
+    public FeatureTableData buildTableDataAndClose(FeatureIndexResults results, double tolerance, LatLng clickLocation, float density, double zoom, View view, GoogleMap map, float screenClickPercentage, Projection projection) {
 
         FeatureTableData tableData = null;
 
         // Fine filter results so that the click location is within the tolerance of each feature row result
-        FeatureIndexResults filteredResults = fineFilterResults(results, tolerance, clickLocation);
+        FeatureIndexResults filteredResults = fineFilterResults(results, tolerance, clickLocation, density, zoom, view, map, screenClickPercentage);
 
         long featureCount = filteredResults.count();
         if (featureCount > 0) {
@@ -600,12 +744,17 @@ public class FeatureInfoBuilder {
     /**
      * Fine filter the index results verifying the click location is within the tolerance of each feature row
      *
-     * @param results       feature index results
-     * @param tolerance     distance tolerance
-     * @param clickLocation click location
+     * @param results               feature index results
+     * @param tolerance             distance tolerance
+     * @param clickLocation         click location
+     * @param density               display density: {@link android.util.DisplayMetrics#density}
+     * @param zoom                  current zoom level
+     * @param view                  view
+     * @param map                   Google Map
+     * @param screenClickPercentage screen click percentage between 0.0 and 1.0
      * @return filtered feature index results
      */
-    private FeatureIndexResults fineFilterResults(FeatureIndexResults results, double tolerance, LatLng clickLocation) {
+    private FeatureIndexResults fineFilterResults(FeatureIndexResults results, double tolerance, LatLng clickLocation, float density, double zoom, View view, GoogleMap map, float screenClickPercentage) {
 
         FeatureIndexResults filteredResults = null;
         if (ignoreGeometryTypes.contains(geometryType)) {
@@ -628,15 +777,21 @@ public class FeatureInfoBuilder {
 
                         if (!ignoreGeometryTypes.contains(geometry.getGeometryType())) {
 
+                            boolean addRow = true;
+
                             if (clickLocation != null) {
 
                                 GoogleMapShape mapShape = converter.toShape(geometry);
-                                if (MapUtils.isPointOnShape(clickLocation, mapShape, geodesic, tolerance)) {
-
-                                    filteredListResults.addRow(featureRow);
-
+                                Boolean styleFiltered = fineFilterStyle(featureRow, geometry, mapShape, clickLocation, density, zoom, view, map, screenClickPercentage);
+                                if (styleFiltered == null) {
+                                    addRow = MapUtils.isPointOnShape(clickLocation, mapShape, geodesic, tolerance);
+                                } else {
+                                    addRow = styleFiltered;
                                 }
-                            } else {
+
+                            }
+
+                            if (addRow) {
                                 filteredListResults.addRow(featureRow);
                             }
 
@@ -650,6 +805,69 @@ public class FeatureInfoBuilder {
         }
 
         return filteredResults;
+    }
+
+    /**
+     * Fine filter the feature row with feature styles
+     *
+     * @param featureRow            feature row
+     * @param geometry              geometry
+     * @param mapShape              Google Map Shape
+     * @param clickLocation         click location
+     * @param density               display density: {@link android.util.DisplayMetrics#density}
+     * @param zoom                  current zoom level
+     * @param view                  view
+     * @param map                   Google Map
+     * @param screenClickPercentage screen click percentage between 0.0 and 1.0
+     * @return true if passes fine filter
+     */
+    private Boolean fineFilterStyle(FeatureRow featureRow, Geometry geometry, GoogleMapShape mapShape, LatLng clickLocation, float density, double zoom, View view, GoogleMap map, float screenClickPercentage) {
+        Boolean passes = null;
+        if (featureStyles != null && view != null && map != null) {
+
+            PixelBounds pixelBounds = null;
+
+            IconRow iconRow = featureStyles.getIcon(featureRow);
+            if (iconRow != null) {
+
+                pixelBounds = FeatureStyleExtension.calculatePixelBounds(iconRow, density);
+
+            } else {
+                StyleRow styleRow = featureStyles.getStyle(featureRow);
+                if (styleRow != null) {
+                    pixelBounds = FeatureStyleExtension.calculatePixelBounds(styleRow, density);
+                }
+            }
+
+            if (pixelBounds != null) {
+
+                // Clear expanded pixel bounds in the click direction opposite of a point
+                if (geometry.getGeometryType() == GeometryType.POINT) {
+                    Point point = (Point) geometry;
+                    if (point.getX() < clickLocation.longitude) {
+                        pixelBounds.setRight(0);
+                    } else if (point.getX() > clickLocation.longitude) {
+                        pixelBounds.setLeft(0);
+                    }
+                    if (point.getY() < clickLocation.latitude) {
+                        pixelBounds.setUp(0);
+                    } else if (point.getY() > clickLocation.latitude) {
+                        pixelBounds.setDown(0);
+                    }
+                    if (pixelBounds.getArea() == 0) {
+                        pixelBounds = null;
+                    }
+                }
+
+                // Get the map click distance tolerance
+                double tolerance = MapUtils.getToleranceDistance(clickLocation, density, zoom, pixelBounds, view, map, screenClickPercentage);
+
+                passes = MapUtils.isPointOnShape(clickLocation, mapShape, geodesic, tolerance);
+
+            }
+
+        }
+        return passes;
     }
 
 }